added fused_moe
